Go routines/channels added for db queries, functions segregation
